Split mobile Track vs Collection Details Tiles
